+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - No longer tested on NodeJS &lt; 16.3.0 and NPM &lt; 7.15.1
 
+## DEPENDENCIES
+
+- **shortid** has been replaced by nanoid
+
 # 5.0.1 / 2020-12-11
 
 ## BUG FIXES